+++ src/components/empty-component/empty-component.tsx
@@ -1,4 +1,4 @@
-import { City } from '../types/city';
+import { City } from '../../types/city';
 
 type EmptyComponentProps = {
   city: City;
