+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/app';
+import App from './components/app/app';
 import store from './store';
 import { checkAuthAction, fetchFavoriteOffers, fetchOffersAction } from './store/api-actions';
 import { Provider } from 'react-redux';
