+++ src/utils.ts
@@ -1,6 +1,7 @@
-import { SortType } from './const';
+import { SortType, cities } from './const';
 import store from './store';
 import { changeFavorite } from './store/api-actions';
+import { City } from './types/city';
 import { Offer } from './types/offer';
 
 
@@ -37,4 +38,8 @@ export function formatDateForView(dateString: string) {
 export function changeFavoriteStatus(offer: Offer, isFavorite: boolean) {
   const status = isFavorite ? 0 : 1;
   store.dispatch(changeFavorite({id: offer.id, status: status}));
+}
+
+export function getRandomCity(): City {
+  return cities[Math.round(Math.random() * 5)];
 }
